+++ tasks.py
@@ -26,9 +26,11 @@ ns.configure(
         "app": "myproject",
         "aws": {
             "region": "us-west-2",
+            "profile_name": "my-aws-profile",  # a profile from .aws/credentials
         },
         "cluster": "Myproject-EKS-cluster",
         "repository": "123456789012.dkr.ecr.us-east-1.amazonaws.com/myproject",
+        "hosting_services_backup_folder": "myproject",
         "run": {
             "echo": True,
             "pty": True,